Fix wrong size_t* to int* conversion in 64bit Big-Endian hosts
- Fixes #636
- Ensure that we have RTP or RTCP packet length into a `int` variable since `srtp_(un)protect_xxxx()` methods expect a pointer to `int` and modify the length referenced by the given pointer.

+++ worker/include/RTC/SrtpSession.hpp
@@ -36,10 +36,10 @@ namespace RTC
 		~SrtpSession();
 
 	public:
-		bool EncryptRtp(const uint8_t** data, size_t* len);
-		bool DecryptSrtp(uint8_t* data, size_t* len);
-		bool EncryptRtcp(const uint8_t** data, size_t* len);
-		bool DecryptSrtcp(uint8_t* data, size_t* len);
+		bool EncryptRtp(const uint8_t** data, int* len);
+		bool DecryptSrtp(uint8_t* data, int* len);
+		bool EncryptRtcp(const uint8_t** data, int* len);
+		bool DecryptSrtcp(uint8_t* data, int* len);
 		void RemoveStream(uint32_t ssrc)
 		{
 			srtp_remove_stream(this->session, uint32_t{ htonl(ssrc) });
